move repo package into model
