+++ app/src/main/AndroidManifest.xml
@@ -32,7 +32,7 @@
         </activity>
         <activity android:name=".ui.manipulating.ManipulatingActivity" />
 
-        <service android:name=".service.ImageObserver" />
+        <service android:name=".model.ImageObserver" />
 
         <receiver
                 android:name=".receiver.BootReceiver"
