+++ app/src/main/java/black/bracken/picsorter/model/ImageObserver.kt
@@ -1,13 +1,13 @@
-package black.bracken.picsorter.service
+package black.bracken.picsorter.model
 
 import android.app.Service
 import android.content.Intent
 import android.os.FileObserver
 import android.os.IBinder
 import black.bracken.picsorter.ext.notificationManager
-import black.bracken.picsorter.notification.DetectionNotification
-import black.bracken.picsorter.notification.ObservingNotification
-import black.bracken.picsorter.service.repository.settings.SettingsRepository
+import black.bracken.picsorter.model.notification.DetectionNotification
+import black.bracken.picsorter.model.notification.ObservingNotification
+import black.bracken.picsorter.model.repository.settings.SettingsRepository
 import kotlinx.coroutines.*
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.flow.Flow
