+++ app/src/main/java/black/bracken/picsorter/ui/dirchooser/DirectoriesChooserViewModel.kt
@@ -3,7 +3,7 @@ package black.bracken.picsorter.ui.dirchooser
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import black.bracken.picsorter.service.repository.settings.SettingsRepository
+import black.bracken.picsorter.model.repository.settings.SettingsRepository
 
 class DirectoriesChooserViewModel(
     private val settingsRepository: SettingsRepository
